[REFACTOR] make the imports more uniform and simpler
