+++ __tests__/pond/[id]/edit.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen, fireEvent, waitFor, act } from '@testing-library/react';
-import EditPond from '@/components/pond/EditPond'; // Adjust the import path as necessary
+import { EditPond } from '@/components/pond';
 import { updatePond } from '@/lib/pond';
 
 jest.mock('@/lib/pond', () => ({
