+++ components/pond/EditPond.tsx
@@ -1,13 +1,13 @@
 'use client';
 
 import React, { useState } from 'react';
-import { Button } from '../ui/button';
+import { Button } from '@/components/ui/button';
 import { updatePond } from '@/lib/pond';
 import { IoIosSave } from 'react-icons/io';
-import { Modal } from '../ui/modal';
+import { Modal } from '@/components/ui/modal';
 import { Pond } from '@/types/pond';
-import PondForm from './PondForm';
-import { PondInputForm } from '@/types/pond/inputpond';
+import { PondForm } from '@/components/pond';
+import { PondInputForm } from '@/types/pond';
 
 interface EditPondProps extends React.HTMLProps<HTMLDivElement> {
   token?: string;
