+++ components/pond/index.tsx
@@ -3,3 +3,4 @@ export { default as PondCard } from '@/components/pond/PondCard';
 export { default as PondList } from '@/components/pond/PondList';
 export { default as DeletePond } from '@/components/pond/DeletePond';
 export { default as EditPond } from '@/components/pond/EditPond';
+export { default as PondForm } from '@/components/pond/PondForm';
